Move testcontainers folder to ./setup, per other projects
