+++ e2e-tests/setup/global-setup.ts
@@ -1,10 +1,10 @@
 /**
  * Playwright global setup. Attaches to a running Grid stack if there is one, otherwise
- * boots a fresh one (see `testcontainers/stack.ts`), stashing it for `global-teardown.ts`
+ * boots a fresh one (see `setup/stack.ts`), stashing it for `global-teardown.ts`
  * and exposing the Kahuna base URL via `GRID_BASE_URL`.
  */
-import { ensureStack } from './testcontainers/stack.ts';
-import { setEnvironment } from './testcontainers/state.ts';
+import { ensureStack } from './stack.ts';
+import { setEnvironment } from './state.ts';
 
 async function globalSetup(): Promise<void> {
   const environment = await ensureStack();
