+++ e2e-tests/setup/global-teardown.ts
@@ -1,8 +1,8 @@
 /**
  * Playwright global teardown. Stops the stack started in `global-setup.ts`.
  */
-import { stopStack } from './testcontainers/stack.ts';
-import { getEnvironment } from './testcontainers/state.ts';
+import { stopStack } from './stack.ts';
+import { getEnvironment } from './state.ts';
 
 async function globalTeardown(): Promise<void> {
   await stopStack(getEnvironment());
