+++ e2e-tests/steps/global-setup.ts
@@ -1,0 +1,15 @@
+/**
+ * Playwright global setup. Attaches to a running Grid stack if there is one, otherwise
+ * boots a fresh one (see `setup/stack.ts`), stashing it for `global-teardown.ts`
+ * and exposing the Kahuna base URL via `GRID_BASE_URL`.
+ */
+import { ensureStack } from '../setup/stack.ts';
+import { setEnvironment } from '../setup/state.ts';
+
+async function globalSetup(): Promise<void> {
+  const environment = await ensureStack();
+  process.env.GRID_BASE_URL = environment.baseUrl;
+  setEnvironment(environment);
+}
+
+export default globalSetup;
